Extract KMS utils into a separate file
